Mapper 수정, Like 합계, 신고하기  6/17 17:38

+++ src/main/java/com/example/lit/domain/dao/review/LikeDAO.java
@@ -15,5 +15,5 @@ public class LikeDAO {
     // 좋아요 취소
     public void remove(LikeVO likeVO){ likeMapper.delete(likeVO);}
     // 좋아요 갯수
-    public int getTotal(Long reviewNumber){ return likeMapper.getTotal(reviewNumber);}
+    public Long getTotal(Long reviewNumber){ return likeMapper.getTotal(reviewNumber);}
 }
